osx in progress

+++ Engine/gapi/shaderreflection.h
@@ -33,7 +33,7 @@ class ShaderReflection final {
       uint32_t layout=0;
       Class    cls   =Ubo;
       Stage    stage =Fragment;
-      uint64_t size  =0;
+      size_t   size  =0;
 
       spirv_cross::ID spvId;
       uint32_t        mslBinding  = uint32_t(-1);
